Fix files impacted by the move
